feat: Restart Navie automatically when flipping VSCode LM option
